db operations

+++ src/models/loan_applications.rs
@@ -2,12 +2,12 @@ use crate::schema;
 use diesel::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Queryable, Selectable, Serialize, Deserialize)]
+#[derive(Queryable, Debug, Selectable, Serialize, Deserialize)]
 #[diesel(table_name = schema::loan_applications)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct LoanApplication {
     pub id: i32,
-    pub amount: f32,
+    pub amount: f64,
     pub borrower_id: i32,
     pub loan_id: i32,
     pub updated_at: chrono::NaiveDateTime,
@@ -18,7 +18,7 @@ pub struct LoanApplication {
 #[diesel(table_name = schema::loan_applications)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct NewLoanApplication {
-    pub amount: f32,
+    pub amount: f64,
     pub borrower_id: i32,
     pub loan_id: i32,
 }
